# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mobile Shop POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
        - name: low_stock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Get product by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CUSTOMERS
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List purchases
      parameters:
        - name: supplier_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Purchase list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseListResponse"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    delete:
      operationId: deletePurchase
      tags: [purchases]
      summary: Delete purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # EXPENSES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Product:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        imei:
          type: ["string", "null"]
        purchase_price:
          type: number
        sale_price:
          type: number
        stock_qty:
          type: integer
        supplier_id:
          type: ["integer", "null"]
        supplier_name:
          type: ["string", "null"]
        warranty:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        low_stock_threshold:
          type: integer
        created_at:
          type: string
      required: [id, name, purchase_price, sale_price, stock_qty]
    ProductInput:
      type: object
      properties:
        barcode:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        color:
          type: ["string", "null"]
        imei:
          type: ["string", "null"]
        purchase_price:
          type: number
        sale_price:
          type: number
        stock_qty:
          type: integer
        supplier_id:
          type: ["integer", "null"]
        warranty:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        low_stock_threshold:
          type: integer
      required: [name, purchase_price, sale_price, stock_qty]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]

    # CUSTOMERS
...
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        unit_cost:
          type: number
        total:
          type: number
      required: [id, product_id, product_name, quantity, unit_cost, total]
    PurchaseItemInput:
      type: object
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
        unit_cost:
          type: number
      required: [product_id, quantity, unit_cost]
    Purchase:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        supplier_id:
          type: ["integer", "null"]
        supplier_name:
          type: ["string", "null"]
        total_amount:
          type: number
        paid_amount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        created_at:
          type: string
      required: [id, invoice_no, total_amount, paid_amount]
    PurchaseInput:
      type: object
      properties:
        supplier_id:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
        paid_amount:
          type: number
        notes:
          type: ["string", "null"]
      required: [items, paid_amount]
    PurchaseListResponse:
      type: object
      properties:
        purchases:
          type: array
          items:
            $ref: "#/components/schemas/Purchase"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [purchases, total, page, limit]

    # EXPENSES
...