# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop POS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @nullable */
  barcode?: string | null;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  category_id?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  imei?: string | null;
  purchase_price: number;
  sale_price: number;
  stock_qty: number;
  /** @nullable */
  supplier_id?: number | null;
  /** @nullable */
  warranty?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  notes?: string | null;
  low_stock_threshold?: number;
}